Update coordinate manager.

+++ cloud/src/main/java/dataelaboration/model/Coordinate.java
@@ -17,7 +17,7 @@ public class Coordinate {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
@@ -25,7 +25,7 @@ public class Coordinate {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
